frontend/requirements.txt
torch
torchvision
opencv-python
gradio
joblib
scikit-learn
imbalanced-learn
numpy
Pillow